+++ gradle-plugin/openshift/src/test/java/org/eclipse/jkube/gradle/plugin/OpenShiftPluginRegisterTaskTest.java
@@ -17,14 +17,13 @@ import java.util.stream.Stream;
 
 import org.eclipse.jkube.gradle.plugin.task.KubernetesConfigViewTask;
 import org.eclipse.jkube.gradle.plugin.task.KubernetesLogTask;
-import org.eclipse.jkube.gradle.plugin.task.KubernetesRemoteDevTask;
-import org.eclipse.jkube.gradle.plugin.task.KubernetesRemoteDevTask;
 import org.eclipse.jkube.gradle.plugin.task.OpenShiftApplyTask;
 import org.eclipse.jkube.gradle.plugin.task.OpenShiftBuildTask;
 import org.eclipse.jkube.gradle.plugin.task.OpenShiftDebugTask;
 import org.eclipse.jkube.gradle.plugin.task.OpenShiftHelmPushTask;
 import org.eclipse.jkube.gradle.plugin.task.OpenShiftHelmTask;
 import org.eclipse.jkube.gradle.plugin.task.OpenShiftPushTask;
+import org.eclipse.jkube.gradle.plugin.task.OpenShiftRemoteDevTask;
 import org.eclipse.jkube.gradle.plugin.task.OpenShiftResourceTask;
 import org.eclipse.jkube.gradle.plugin.task.OpenShiftUndeployTask;
 
@@ -56,7 +55,7 @@ class OpenShiftPluginRegisterTaskTest {
         arguments("ocUndeploy", OpenShiftUndeployTask.class),
         arguments("ocHelm", OpenShiftHelmTask.class),
         arguments("ocHelmPush", OpenShiftHelmPushTask.class),
-        arguments("ocRemoteDev", KubernetesRemoteDevTask.class),
+        arguments("ocRemoteDev", OpenShiftRemoteDevTask.class),
         arguments("ocWatch", OpenShiftWatchTask.class));
   }