+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Usage:
 ```
 ### 1.11-SNAPSHOT
 * Fix #1546: Migrate to JUnit5 testing framework
+* Fix #1935: `oc:remote-dev` goal / `ocRemoteDev` task have wrong log prefixes
 
 ### 1.10.1 (2022-11-16)
 * Fix #1915: Maven 3.6.3 is still supported